pyproject.toml
[project]
name = "asparagus"
version = "0.3.0"
authors = [
  { name="Sebastian Llambias", email="llambias@live.com" },
]
description = "Yucca-spinoff for foundation model training"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License 2.0",
    "Operating System :: OS Independent",
]
keywords = ['deep learning', 'medical image analysis','foundation models']


dependencies = [
    "gardening_tools>=0.1.1",
    "lightning==2.4.0",
    "nibabel>=5.3.2",
    "numpy>=1.23.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "Pillow>=12.0.0",
    "python-dotenv>=1.0.0",
    "torch==2.6.0",
    "torchmetrics>=1.7.2",
    "torchvision==0.21.0",
    "wandb>=0.23.0",
]


[dependency-groups]
dcai = [
    "numpy<2.0.0",
    "torch==2.6.0",
    "hydra-core>=1.3.2",
    "torchmetrics>=1.7.2",
    "lightning==2.4.0",
    "pytorch-lightning==2.5.0",
    "monai==1.4.0",
    "pandas==2.2.2",
    "wandb==0.21.3",
]


[project.optional-dependencies]
extras = [
    "mlflow>=3.0.0",
    "gardening_tools[extras]>=0.1.1",    
]
test = [
    "ruff>=0.14.8",
    "pytest>=9.0.1"
]


[project.scripts]
asp_getid = 'asparagus.pipeline.run.get_id:main'
asp_pretrain = 'asparagus.pipeline.run.pretrain:main'
asp_train_seg = 'asparagus.pipeline.run.train_seg:main'
asp_test_seg = 'asparagus.pipeline.run.test_seg:main'
asp_train_cls = 'asparagus.pipeline.run.train_cls:main'
asp_train_reg = 'asparagus.pipeline.run.train_reg:main'
asp_finetune_seg = 'asparagus.pipeline.run.finetune_seg:main'
asp_finetune_cls = 'asparagus.pipeline.run.finetune_cls:main'
asp_finetune_reg = 'asparagus.pipeline.run.finetune_reg:main'
asp_eval_box_run = 'asparagus.pipeline.run.eval_box:main'
asp_eval_box_prepare_data = 'asparagus.pipeline.run.eval_box:prepare_data'
asp_eval_box_collect_results = 'asparagus.pipeline.run.eval_box:collect_results'


[tool.setuptools.packages.find]
include = ["asparagus*"]


[project.urls]
"Homepage" = "https://github.com/sllambias/asparagus"


[build-system]
requires = ["setuptools>=75.8", "wheel>=0.45"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 127


[tool.ruff.lint]
# "E": style/error codes (PEP 8–type errors, e.g. if x == None).
# "F": Pyflakes-style errors (unused variables, undefined names, etc.).
# "I": import-related issues (import order, duplicates, etc.).
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
no-sections = true
combine-as-imports = true

[tool.ruff.format]
# quote-style = "double": Ruff prefers " over ' when it reformats strings, as long as it’s safe.
# indent-style = "space": indentation uses spaces, not tabs.
quote-style = "double"
indent-style = "space"


[tool.pytest.ini_options]
# addopts = "-ra -q" adds default CLI options:
#     -q: quiet mode (less verbose).
#     -ra: show a short summary of skipped/failed/xpassed tests at the end.
# testpaths = ["tests"]: when you run pytest with no args, it looks for tests in the tests/ directory.
addopts = "-ra -q"
testpaths = ["tests"]

